requirements.txt
torchvision
torch
spacy
pillow
matplotlib